requirements.txt
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
drf-yasg==1.21.9
envparse==0.2.0
pillow==11.1.0
psycopg2-binary==2.9.9
